# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: ms-case-data
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "8.0"
servers:
  - url: http://localhost:64895
    description: Generated server url
...
paths:
...
  /{municipalityId}/notes/{noteId}:
    get:
      tags:
        - Notes
      description: Get note by id
      operationId: getNoteById
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDTO"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/NoteDTO"
    delete:
      tags:
        - Notes
      description: Delete a note by note id
      operationId: deleteNoteById
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No content - Successful operation
        "403":
          description: Forbidden
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
        - Notes
      description: Update note
      operationId: patchNoteOnErrand
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteDTO"
        required: true
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: No content - Successful operation
...
  /{municipalityId}/parking-permits:
    get:
      tags:
        - Parking permit
      description: Get all issued parking permits.
      operationId: getAllParkingPermits
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: personId
          in: query
          description: personId of the applicant
          required: false
          schema:
            type: string
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetParkingPermitDTO"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetParkingPermitDTO"
...
  /{municipalityId}/notes/errand/{errandId}:
    get:
      tags:
        - Notes
      description: "Get notes for a specific errand, possible to filter by note type"
      operationId: getNotesByErrandId
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: errandId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: noteType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NoteType"
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteDTO"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteDTO"
...
  /{municipalityId}/errands/{errandId}/notes/{noteId}:
    delete:
      tags:
        - Errands
      description: Delete note on errand.
      operationId: deleteNote
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: errandId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: No content - Successful operation
...
components:
  schemas:
    AddressDTO:
      type: object
      properties:
        addressCategory:
          type: string
          enum:
            - POSTAL_ADDRESS
            - INVOICE_ADDRESS
            - VISITING_ADDRESS
        street:
          maxLength: 255
          minLength: 0
          type: string
          example: Testvägen
        houseNumber:
          maxLength: 255
          minLength: 0
          type: string
          example: "18"
        postalCode:
          maxLength: 255
          minLength: 0
          type: string
          example: 123 45
        city:
          maxLength: 255
          minLength: 0
          type: string
          example: Sundsvall
        country:
          maxLength: 255
          minLength: 0
          type: string
          example: Sverige
        careOf:
          maxLength: 255
          minLength: 0
          type: string
          description: c/o
          example: Test Testorsson
        attention:
          maxLength: 255
          minLength: 0
          type: string
          example: Test Testorsson
        propertyDesignation:
          maxLength: 255
          minLength: 0
          type: string
          example: SUNDSVALL LJUSTA 7:2
        apartmentNumber:
          maxLength: 255
          minLength: 0
          type: string
          example: LGH 1001
        isZoningPlanArea:
          type: boolean
        invoiceMarking:
          maxLength: 255
          minLength: 0
          type: string
          description: "Only in combination with addressCategory: INVOICE_ADDRESS"
        location:
          $ref: "#/components/schemas/CoordinatesDTO"
      description: An stakeholder may have one or more addresses. For example one
        POSTAL_ADDRESS and another INVOICE_ADDRESS.
    ContactInformationDTO:
      type: object
      properties:
        contactType:
          type: string
          enum:
            - CELLPHONE
            - PHONE
            - EMAIL
        value:
          maxLength: 255
          minLength: 0
          type: string
    CoordinatesDTO:
      type: object
      properties:
        latitude:
          type: number
          description: Decimal Degrees (DD)
          format: double
          example: 62.390205
        longitude:
          type: number
          description: Decimal Degrees (DD)
          format: double
          example: 17.306616
    StakeholderDTO:
      required:
        - roles
        - type
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        version:
          type: integer
          format: int32
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        type:
          type: string
          enum:
            - PERSON
            - ORGANIZATION
        firstName:
          maxLength: 255
          minLength: 0
          type: string
          example: Test
        lastName:
          maxLength: 255
          minLength: 0
          type: string
          example: Testorsson
        personId:
          type: string
          example: 3ed5bc30-6308-4fd5-a5a7-78d7f96f4438
        organizationName:
          maxLength: 255
          minLength: 0
          type: string
          example: Sundsvalls testfabrik
        organizationNumber:
          maxLength: 255
          minLength: 0
          pattern: "^((18|19|20|21)\\d{6}|\\d{6})-(\\d{4})$"
          type: string
          description: Organization number with 10 or 12 digits.
          example: 19901010-1234
        authorizedSignatory:
          maxLength: 255
          minLength: 0
          type: string
          example: Test Testorsson
        adAccount:
          maxLength: 36
          minLength: 0
          type: string
          description: AD-account
        roles:
          type: array
          description: An stakeholder can have one or more roles.
          items:
            type: string
            description: An stakeholder can have one or more roles.
        addresses:
          type: array
          description: An stakeholder may have one or more addresses. For example
            one POSTAL_ADDRESS and another INVOICE_ADDRESS.
          items:
            $ref: "#/components/schemas/AddressDTO"
        contactInformation:
          type: array
          items:
            $ref: "#/components/schemas/ContactInformationDTO"
        extraParameters:
          type: object
          additionalProperties:
            type: string
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    AttachmentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        version:
          type: integer
          format: int32
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        category:
          type: string
        name:
          type: string
        note:
          type: string
        extension:
          type: string
        mimeType:
          type: string
        file:
          type: string
        errandNumber:
          type: string
        extraParameters:
          type: object
          additionalProperties:
            type: string
    DecisionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        version:
          type: integer
          format: int32
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        decisionType:
          type: string
          enum:
            - RECOMMENDED
            - PROPOSED
            - FINAL
        decisionOutcome:
          type: string
          enum:
            - APPROVAL
            - REJECTION
            - DISMISSAL
            - CANCELLATION
        description:
          maxLength: 100000
          minLength: 0
          type: string
        law:
          type: array
          items:
            $ref: "#/components/schemas/LawDTO"
        decidedBy:
          $ref: "#/components/schemas/StakeholderDTO"
        decidedAt:
          type: string
          format: date-time
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentDTO"
        extraParameters:
          type: object
          additionalProperties:
            type: string
    LawDTO:
      type: object
      properties:
        heading:
          maxLength: 255
          minLength: 0
          type: string
        sfs:
          maxLength: 255
          minLength: 0
          type: string
        chapter:
          maxLength: 255
          minLength: 0
          type: string
        article:
          maxLength: 255
          minLength: 0
          type: string
...
    NoteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        version:
          type: integer
          format: int32
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        title:
          maxLength: 255
          minLength: 0
          type: string
          example: Motivering till bifall
        text:
          maxLength: 10000
          minLength: 0
          type: string
          example: Den sökande har rätt till parkeringstillstånd eftersom alla kriterier
            uppfylls.
        createdBy:
          maxLength: 36
          minLength: 0
          type: string
          description: AD-account for the user who created the note.
          readOnly: true
          example: user
        updatedBy:
          maxLength: 36
          minLength: 0
          type: string
          description: AD-account for the user who last modified the note.
          readOnly: true
          example: user
        noteType:
          $ref: "#/components/schemas/NoteType"
        extraParameters:
          type: object
          additionalProperties:
            type: string
    NoteType:
      type: string
      description: The type of note
      example: INTERNAL
      enum:
        - INTERNAL
        - PUBLIC
...
    GetParkingPermitDTO:
      type: object
      properties:
        artefactPermitNumber:
          type: string
        artefactPermitStatus:
          type: string
        errandId:
          type: integer
          format: int64
        errandDecision:
          $ref: "#/components/schemas/DecisionDTO"
...
  securitySchemes: { }